rfmp-daemon/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async and networking
asyncio-mqtt==0.16.1
pyserial-asyncio==0.6

# Database
aiosqlite==0.19.0

# Hashing and crypto
mmh3==4.0.1

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
python-multipart==0.0.6

# Logging
structlog==23.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.0